third_party/blink/web_tests/platform/generic/http/tests/websocket/workers/worker-simple-expected.txt

Test for WebSocket in Worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS: worker: init
PASS: worker: WebSocket exists
PASS: worker: successfullyParsed:[object WebSocket]
PASS: worker: Connected.
PASS: worker: Received: 'Hello from Simple WSH.'
PASS: worker: Closed.
PASS: worker: closeEvent.wasClean is true.
DONE
TEST COMPLETE
